requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
geopy
requests
pymysql
cryptography